factor PRNG routines from random module to prng
