+++ jax/random.py
@@ -78,8 +78,15 @@ for the design and its motivation.
 
 # flake8: noqa: F401
 
-from jax._src.random import (
+from jax._src.prng import (
   PRNGKey,
+  fold_in,
+  split,
+  threefry2x32_p,
+  threefry_2x32,
+)
+
+from jax._src.random import (
   bernoulli,
   beta,
   categorical,
@@ -88,7 +95,6 @@ from jax._src.random import (
   dirichlet,
   double_sided_maxwell,
   exponential,
-  fold_in,
   gamma,
   gumbel,
   laplace,
@@ -103,10 +109,7 @@ from jax._src.random import (
   randint,
   random_gamma_p,
   shuffle,
-  split,
   t,
-  threefry2x32_p,
-  threefry_2x32,
   truncated_normal,
   uniform,
   weibull_min,
